python/shared/pipecat/pyproject.toml
[tool.poetry]
name = "pipecat"
version = "0.1.0"
description = "Pipecat implementation"
authors = ["Andrew Chang <andrew@tryelixir.io>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anthropic = "0.25.9"
anyio = "4.3.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
av = "12.0.0"
azure-cognitiveservices-speech = "1.37.0"
blinker = "1.8.2"
cachetools = "5.3.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
coloredlogs = "15.0.1"
ctranslate2 = "4.2.1"
daily-python = "0.8.0"
distro = "1.9.0"
einops = "0.8.0"
exceptiongroup = "1.2.1"
fal-client = "0.4.0"
faster-whisper = "1.0.2"
filelock = "3.14.0"
flask = "3.0.3"
flask-cors = "4.0.1"
flatbuffers = "24.3.25"
frozenlist = "1.4.1"
fsspec = "2024.5.0"
future = "1.0.0"
google-ai-generativelanguage = "0.6.4"
google-api-core = {version = "2.19.0", extras = ["grpc"]}
google-api-python-client = "2.130.0"
google-auth = "2.29.0"
google-auth-httplib2 = "0.2.0"
google-generativeai = "0.5.4"
googleapis-common-protos = "1.63.0"
grpcio = "1.64.0"
grpcio-status = "1.62.2"
h11 = "0.14.0"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httpx = "0.27.0"
httpx-sse = "0.4.0"
huggingface-hub = "0.23.1"
humanfriendly = "10.0"
idna = "3.7"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
loguru = "0.7.2"
markupsafe = "2.1.5"
mpmath = "1.3.0"
multidict = "6.0.5"
networkx = "3.3"
numpy = "1.26.4"
onnxruntime = "1.18.0"
openai = "1.26.0"
packaging = "24.0"
pillow = "10.3.0"
proto-plus = "1.23.0"
protobuf = "4.25.3"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pyaudio = "0.2.14"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pyht = "0.0.28"
pyloudnorm = "0.1.1"
pyparsing = "3.1.2"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
regex = "2024.5.15"
requests = "2.32.2"
rsa = "4.9"
safetensors = "0.4.3"
scipy = "1.13.1"
sniffio = "1.3.1"
sympy = "1.12"
timm = "0.9.16"
tokenizers = "0.19.1"
torch = "2.3.0"
torchaudio = "2.3.0"
torchvision = "0.18.0"
tqdm = "4.66.4"
transformers = "4.40.2"
typing-extensions = "4.11.0"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
websockets = "12.0"
werkzeug = "3.0.3"
yarl = "1.9.4"
opentelemetry-sdk = "^1.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"